requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
numpy>=1.26.0
pydantic>=2.5.0
pandas>=2.0.0
azure-identity>=1.15.0
pyodbc>=4.0.39
sqlalchemy>=2.0.0
